rename androidtestorchestrator -> mobiletestorchestrator
