+++ orchestrator/test/conftest.py
@@ -9,9 +9,9 @@ import pytest
 from pytest_mproc.plugin import TmpDirFactory
 from typing import Optional
 
-from androidtestorchestrator.application import Application, TestApplication, ServiceApplication
-from androidtestorchestrator.device import Device
-from androidtestorchestrator.emulators import EmulatorBundleConfiguration, Emulator
+from mobiletestorchestrator.application import Application, TestApplication, ServiceApplication
+from mobiletestorchestrator.device import Device
+from mobiletestorchestrator.emulators import EmulatorBundleConfiguration, Emulator
 from . import support
 from .support import uninstall_apk, find_sdk
