+++ orchestrator/test/test_instrumentation_output_parser_class.py
@@ -2,11 +2,11 @@ from typing import Mapping, Any
 
 import os
 
-from androidtestorchestrator.device import Device
-from androidtestorchestrator.devicelog import DeviceLog
-from androidtestorchestrator.parsing import InstrumentationOutputParser
-from androidtestorchestrator.reporting import TestRunListener
-from androidtestorchestrator.timing import StopWatch
+from mobiletestorchestrator.device import Device
+from mobiletestorchestrator.devicelog import DeviceLog
+from mobiletestorchestrator.parsing import InstrumentationOutputParser
+from mobiletestorchestrator.reporting import TestRunListener
+from mobiletestorchestrator.timing import StopWatch
 
 
 class TestInstrumentationOutputParser(object):
