+++ orchestrator/test/test_upgrade_runner_class.py
@@ -3,9 +3,9 @@ from contextlib import suppress
 
 import pytest
 
-from androidtestorchestrator import Device
-from androidtestorchestrator.reporting import TestRunResult
-from androidtestorchestrator.runners.upgradetestrunner import UpgradeTestException, UpgradeTestRunner
+from mobiletestorchestrator import Device
+from mobiletestorchestrator.reporting import TestRunResult
+from mobiletestorchestrator.runners.upgradetestrunner import UpgradeTestException, UpgradeTestRunner
 
 reporter = TestRunResult()
